AVRO-2199: Code review comments
